pyproject.toml
[project]
name = "simple-coding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "e2b-code-interpreter>=2.4.1",
    "openai>=2.9.0",
    "python-dotenv>=1.2.1",
]
